Refactor project structure to TypeScript modules and remove vendored deep-diff
- Replace vendored deep-diff UMD module with modular TypeScript implementation
- Split deep-diff functionality into separate files: apply, calculate, changes, hash, index, types, utils
- Move identity-related functions (getId, indexById, findRemoved) to identity.ts
- Move formatUpdated to format-updated.ts
- Update classify.ts to use new modular structure
- Remove deep-diff-types.ts as types moved to deep-diff/types.ts
- Update tsconfig to exclude lib/deep-diff/** from coverage
- Modernize vitest and tsconfig configurations
