+++ pnpm-workspace.yaml
@@ -1,2 +1,2 @@
 allowBuilds:
-  esbuild: set this to true or false
+  esbuild: true
