+++ tsconfig.json
@@ -16,5 +16,5 @@
     "types": ["node"]
   },
   "include": ["lib/**/*.ts"],
-  "exclude": ["node_modules", "dist", "coverage", "lib/deep-diff.ts"]
+  "exclude": ["node_modules", "dist", "coverage"]
 }
