+++ vitest.config.ts
@@ -2,14 +2,15 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
+    globals: true,
     include: ['test/**/*.spec.ts'],
-    // The vendored UMD module has no meaningful coverage value and makes the
-    // v8 provider emit noisy reports.
+    // The vendored module retains its upstream compatibility suite; coverage
+    // here focuses on this package's own implementation.
     coverage: {
       provider: 'v8',
       reporter: ['text', 'html'],
       include: ['lib/**/*.ts'],
-      exclude: ['lib/deep-diff.ts'],
+      exclude: ['lib/deep-diff/**'],
     },
   },
 });
